fix: change json form request api
Just hacking the content type is not sufficient. It is necessary to
encode the body in getBody methods. Look up Request class in Volley, you
will find that it request with application/x-www-form-urlencoded content
type and encode the parameters that will be post. So I just derive
Request class and make a new JsonFormRequest class, which derived the
content type and getBody from Request. Change JsonObjectFormRequest and
JsonArrayFormRequest to extend JsonFormRequest and change constructors