feat: abandon the way of polling for new messages and refresh message when receive notification
Polling is not a good idea for client and it generate lots of useless
netword overhead. Long polling require server to support but I don't
have time to implement it. Using GCM push is a good way other than
no gurantee of 100% received.

Currently, refresh the message when InboxFragment (MainActivity)
resumes. If InboxFragment (MainActivity) is running, refresh message
by notification because a receiver is registered.