fix: upload user profile (especailly for last calendar) when kicked out.
Don't set last calendar in checklogin. If one can login in check login,
it proves that he has never logout so the user profile is synchronous
with that on the server. If one has to login in check login, it proves
that he has logged out and will login manually and get the last user
profile from the server.